add delete bukti_transaksi #27

+++ api/telkom_apps/controller/delete_user.php
@@ -12,11 +12,19 @@ if (isset($_GET)){
     $iduser = $_GET['id_user'];
 
     $queryDelete = mysqli_query($db, "delete from user_apps where id_user = '$iduser';");
-    $queryDeletePoint = mysqli_query($db, "delete from bukti_transaksi_user where id_user = '$iduser';");
+    if ($queryDelete) {
 
-    header('Location: ../../../pages/tables/data_user.php');
+
+        $queryDeletePoint = mysqli_query($db, "delete from bukti_transaksi_user where id_user = '$iduser';");
+
+        header('Location: ../../../pages/tables/data_user.php');
+    } else{
+
+        $response["error"] = "Error deleting user";
+        echo json_encode($response);
+    }
 } else{
-    $response["error"] = "Error deleting user";
+    $response["error"] = "Error deleting not responding";
     echo json_encode($response);
 }
 ?>
